pyproject.toml
[project]
name = "semanticsql"
version = "0.1.0"
description = "SemanticSQL is a FastAPI application that allows users to connect to external databases, extract schema metadata, and use natural language to generate SQL queries. It leverages semantic search and embeddings to understand database schema structure and generate appropriate SQL however complex it is."
authors = [
    {name = "Ashutosh",email = "ashutoshpatra1135@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pydantic (>=2.11.3,<3.0.0)",
    "langchain (>=0.3.24,<0.4.0)",
    "langchain-google-genai (>=2.1.3,<3.0.0)",
    "qdrant-client (>=1.7.0,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "requests (>=2.32.3,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pytest-asyncio (>=0.26.0,<0.27.0)",
    "coverage (>=7.8.0,<8.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "ruff (>=0.11.7,<0.12.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "langchain-core (>=0.3.56,<0.4.0)",
    "langchain-openai (>=0.3.14,<0.4.0)"
]

[tool.poetry]
packages = [{include = "app"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
